feat: enable creators to select redirect url
Currently creators are unable to set the homepage of their site as a template
page e.g. product page, and they are unable to redirect the homepage to another
page. As a workaround we use a location.replace to force the browser to load
their chosen page.

Previously adding the script would create the redirect but break the site
preview, therefore requiring further support from frontend to make any updates.
By introducing a check that the current hostname does not contain
myeasol we can enable creators to edit the block in the site admin and update
the page they are redirecting to without further support from frontend
enginnering.