Add troubleshooting for detail on past/ongoing slots
We get a lot of queries about the exact logic that applies with
regard to this across Easol internally so I can forsee theme
developers getting confused by it too. They should now be able to
find this troubleshooting article when they search 'past' or 'ongoing'.

Rather selfishly I would like to have this published so I can refer
to it when I forget as well.