Document PackageBooking drop
We've just added `adult_count`, `selected_start_date` and
`selected_start_time` to the PackageBooking drop.

The drop did already exist and already had the `id` and `package`
attributes.

I *think* that it was already exposed to the package booking template,
but we're now going to expose it as a `#package_booking` attribute of
CartItem drops, too.