Update variant_pricing tag documentation
We have made two changes to how this tag works this week. More simple
to document is `include_fees` parameter.

Additionally, we ensure that a package step item's adopt the price of
the slot that will be purchased. This is well defined for items with
'exact' or 'closest match' slot selection strateyg. For those  with a
'range' slot selection strategy, the customer will be asked to select
the slot as a following interaction, so we return the cheapest of the
slots they can choose from - we recommend pre-pending the price with
'From' in this case.