chore: ensure docs cover all use-cases
The docs currently do not make it clear that it can be used for fixed-date packages, with the absence of providing a date/time.

Added an example to show how the tag works when you do not pass in any parameters, and updated wording to make it clear that all params are optional.